# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: |
    API specification for WASAText messaging application.
    Built according to the PDF specification.
  version: "1.0.0"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Use the user identifier returned from doLogin
  schemas:
    User:
      type: object
      description: Representation of a user in the system
      properties:
        identifier:
          type: string
          description: Unique user identifier (UUID hex string)
          example: "abcdef012345"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]+$'
        name:
          type: string
          description: Username chosen by the user (3-16 characters)
          minLength: 3
          maxLength: 16
          pattern: '^[a-zA-Z0-9_-]+$'
          example: "Maria"
        photo:
          type: string
          format: binary
          description: User profile photo in binary format (optional)
    Group:
      type: object
      description: Representation of a group conversation
      properties:
        groupId:
          type: string
          description: Unique group identifier
          example: "group123"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]+$'
        name:
          type: string
          description: Name of the group
          example: "Study Group"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9 _-]+$'
        photo:
          type: string
          format: binary
          description: Group photo in binary format (optional)
        members:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: '#/components/schemas/User'
          description: List of users who are members of the group
...
    ConversationPreview:
      type: object
      description: Summary of a conversation for the inbox list view
      properties:
        conversationId:
          type: string
          description: Unique conversation identifier
          example: "conv123"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]+$'
        isGroup:
          type: boolean
          description: Specifies if this conversation is a group chat
          example: false
        name:
          type: string
          description: Display name (username for individual, or group name)
          example: "Maria"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9 _-]+$'
        photo:
          type: string
          format: binary
          description: Profile or group photo binary data
        lastMessageTimestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of the most recent message
          example: "2024-01-15T10:30:00Z"
          minLength: 20
          maxLength: 30
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}'
        lastMessagePreview:
          type: string
          description: Truncated text of the last message
          example: "Hey, how are you?"
          minLength: 0
          maxLength: 100
          pattern: '^[\s\S]*$'
        lastMessageIsPhoto:
          type: boolean
          description: True if the last message in the thread was media
          example: false
    Conversation:
      type: object
      description: Complete details of a conversation thread including all history
      properties:
        conversationId:
          type: string
          description: Unique conversation identifier
          example: "conv123"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]+$'
        isGroup:
          type: boolean
          description: True for group conversations
        name:
          type: string
          description: Name of the conversation
          example: "Maria"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9 _-]+$'
        photo:
          type: string
          format: binary
          description: Image representing the conversation
        members:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: '#/components/schemas/User'
          description: List of users in this conversation (for groups)
        messages:
          type: array
          minItems: 0
          maxItems: 5000
          items:
            $ref: '#/components/schemas/Message'
          description: Ordered list of messages in the conversation (newest first)
    Error:
      type: object
      description: Standard error response object
      properties:
        message:
          type: string
          description: Human-readable explanation of what went wrong
          example: "Username already taken"
          minLength: 1
          maxLength: 256
          pattern: '^[\s\S]+$'
  parameters:
    UserId:
      name: userId
      in: path
      description: Targeted user identifier
      required: true
      schema:
        type: string
        description: User identifier value
        minLength: 1
        maxLength: 64
        pattern: '^[a-zA-Z0-9_-]+$'
    ConversationId:
      name: conversationId
      in: path
      description: Unique conversation identifier
      required: true
      schema:
        type: string
        description: Conversation identifier value
        minLength: 1
        maxLength: 64
        pattern: '^[a-zA-Z0-9_-]+$'
    MessageId:
      name: messageId
      in: path
      description: Unique message identifier
      required: true
      schema:
        type: string
        description: Message identifier value
        minLength: 1
        maxLength: 64
        pattern: '^[a-zA-Z0-9_-]+$'
    GroupId:
      name: groupId
      in: path
      description: Group identifier
      required: true
      schema:
        type: string
        description: Group identifier value
        minLength: 1
        maxLength: 64
        pattern: '^[a-zA-Z0-9_-]+$'
paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User nickname for identification
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Sign-in request body
              properties:
                name:
                  type: string
                  description: Nickname chosen by the user
                  example: Maria
                  pattern: '^[a-zA-Z0-9_-]+$'
                  minLength: 3
                  maxLength: 16
              required:
                - name
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                description: Successful login response
                properties:
                  identifier:
                    type: string
                    description: The unique identifier assigned to the user
                    example: "abcdef012345"
                    minLength: 1
                    maxLength: 64
                    pattern: '^[a-zA-Z0-9_-]+$'
  /users/{userId}/username:
    parameters:
      - $ref: '#/components/parameters/UserId'
    put:
      tags: ["user"]
      summary: Update the user's username
      description: |
        Users can update their name, provided the new name 
        is not already in use by someone else.
      operationId: setMyUserName
      security:
        - bearerAuth: []
      requestBody:
        description: The new nickname to set
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Username update request
              properties:
                name:
                  type: string
                  description: New nickname
                  minLength: 3
                  maxLength: 16
                  pattern: '^[a-zA-Z0-9_-]+$'
                  example: "NewName"
              required:
                - name
      responses:
        '200':
          description: Username updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Empty successful response
                properties:
                  status:
                    type: string
                    description: Operation result status
                    example: "success"
                    minLength: 1
                    maxLength: 16
                    pattern: '^[a-zA-Z]+$'
        '400':
          description: Invalid username format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/photo:
    parameters:
      - $ref: '#/components/parameters/UserId'
    put:
      tags: ["user"]
      summary: Set the user's profile photo
      description: Upload or update the user's profile photo.
      operationId: setMyPhoto
      security:
        - bearerAuth: []
      requestBody:
        description: The image file data
        required: true
        content:
          image/*:
            schema:
              type: string
              format: binary
              description: Binary image data
      responses:
        '200':
          description: Photo updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Empty successful response
                properties:
                  status:
                    type: string
                    description: Operation result status
                    example: "success"
                    minLength: 1
                    maxLength: 16
                    pattern: '^[a-zA-Z]+$'
        '400':
          description: Invalid photo format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users:
    get:
      tags: ["user"]
      summary: Search for users by username
      description: |
        User can search for other users via the username 
        and see all the existing WASAText usernames.
      operationId: searchUsers
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            description: Search query string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]*$'
          description: Username to search for (partial match)
      responses:
        '200':
          description: List of matching users
          content:
            application/json:
              schema:
                type: array
                description: Array of users matching the search
                minItems: 0
                maxItems: 100
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations:
    get:
      tags: ["conversation"]
      summary: Get list of all conversations
      description: |
        Returns a list of conversations with other users or groups,
        sorted in reverse chronological order.
      operationId: getMyConversations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                description: Array of conversation summaries
                minItems: 0
                maxItems: 1000
                items:
                  $ref: '#/components/schemas/ConversationPreview'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: ["conversation"]
      summary: Start a new conversation
      description: |
        The user can start a new conversation with any other user of
        WASAText, and this conversation will automatically be added
        to the list.
      operationId: startConversation
      security:
        - bearerAuth: []
      requestBody:
        description: The user to start a conversation with
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Start conversation request
              properties:
                userId:
                  type: string
                  description: Identifier of the user to start chatting with
                  example: "abcdef012345"
                  minLength: 1
                  maxLength: 64
                  pattern: '^[a-zA-Z0-9_-]+$'
              required:
                - userId
      responses:
        '201':
          description: Conversation created or found successfully
          content:
            application/json:
              schema:
                type: object
                description: Conversation created response
                properties:
                  conversationId:
                    type: string
                    description: Unique identifier of the conversation
                    example: "conv123"
                    minLength: 1
                    maxLength: 64
                    pattern: '^[a-zA-Z0-9_-]+$'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Target user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
    get:
      tags: ["conversation"]
      summary: Get a specific conversation with all messages
      description: |
        Opens a conversation to view all exchanged messages,
        displayed in reverse chronological order.
      operationId: getConversation
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}/messages:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
    post:
      tags: ["message"]
      summary: Send a new message
      description: |
        Send a new message in a conversation. Can be text or photo/GIF.
        Can optionally be a reply to an existing message.
      operationId: sendMessage
      security:
        - bearerAuth: []
      requestBody:
        description: The content of the message
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Text message request
              properties:
                content:
                  type: string
                  description: Text content of the message
                  example: "Hello!"
                  minLength: 1
                  maxLength: 10000
                  pattern: '^[\s\S]+$'
                replyTo:
                  type: string
                  description: Message ID to reply to (optional)
                  example: "msg100"
                  minLength: 1
                  maxLength: 64
                  pattern: '^[a-zA-Z0-9_-]+$'
          multipart/form-data:
            schema:
              type: object
              description: Media message request
              properties:
                photo:
                  type: string
                  format: binary
                  description: Photo or GIF data
                replyTo:
                  type: string
                  description: Message ID to reply to (optional)
                  minLength: 1
                  maxLength: 64
                  pattern: '^[a-zA-Z0-9_-]+$'
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Invalid message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}/messages/{messageId}/forward:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
      - $ref: '#/components/parameters/MessageId'
    post:
      tags: ["message"]
      summary: Forward a message to another conversation
      description: Forward an existing message to another conversation.
      operationId: forwardMessage
      security:
        - bearerAuth: []
      requestBody:
        description: Destination of the forwarded message
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Forward request
              properties:
                targetConversationId:
                  type: string
                  description: Conversation ID where the message should be sent
                  example: "conv456"
                  minLength: 1
                  maxLength: 64
                  pattern: '^[a-zA-Z0-9_-]+$'
              required:
                - targetConversationId
      responses:
        '201':
          description: Message forwarded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation or message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}/messages/{messageId}:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
      - $ref: '#/components/parameters/MessageId'
    delete:
      tags: ["message"]
      summary: Delete a sent message
      description: |
        Delete a message that was sent by the current user.
      operationId: deleteMessage
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Message deleted successfully
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied (not your message)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation or message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}/messages/{messageId}/comments:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
      - $ref: '#/components/parameters/MessageId'
    post:
      tags: ["comment"]
      summary: Add a reaction (comment) to a message
      description: React to a message with an emoticon.
      operationId: commentMessage
      security:
        - bearerAuth: []
      requestBody:
        description: The emoticon for the reaction
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Reaction request body
              properties:
                emoticon:
                  type: string
                  description: Single emoji/emoticon character
                  example: "👍"
                  minLength: 1
                  maxLength: 8
                  pattern: '^[\s\S]+$'
              required:
                - emoticon
      responses:
        '201':
          description: Reaction added successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  status:
                    type: string
                    description: Operation result status
                    example: "success"
                    minLength: 1
                    maxLength: 16
                    pattern: '^[a-zA-Z]+$'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation or message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: ["comment"]
      summary: Remove a reaction (uncomment) from a message
      description: Remove your own reaction from a message.
      operationId: uncommentMessage
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Reaction removed successfully
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Reaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups:
    post:
      tags: ["group"]
      summary: Create a new group
      description: |
        User can create a new group with any number of other WASAText users.
      operationId: createGroup
      security:
        - bearerAuth: []
      requestBody:
        description: Details of the new group and initial members
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Group creation request
              properties:
                name:
                  type: string
                  description: Descriptive name for the group
                  example: "Study Group"
                  minLength: 1
                  maxLength: 64
                  pattern: '^[a-zA-Z0-9 _-]+$'
                memberIds:
                  type: array
                  description: Initial list of member identifiers
                  minItems: 1
                  maxItems: 1000
                  items:
                    type: string
                    description: User identifier
                    minLength: 1
                    maxLength: 64
                    pattern: '^[a-zA-Z0-9_-]+$'
                  example: ["abcdef012345", "fedcba987654"]
              required:
                - name
                - memberIds
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid group data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/members:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    post:
      tags: ["group"]
      summary: Add a user to the group
      description: Group members can add other users to the group.
      operationId: addToGroup
      security:
        - bearerAuth: []
      requestBody:
        description: ID of the user to be added
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Add member request
              properties:
                userId:
                  type: string
                  description: Identifier of the user to invite
                  example: "abcdef012345"
                  minLength: 1
                  maxLength: 64
                  pattern: '^[a-zA-Z0-9_-]+$'
              required:
                - userId
      responses:
        '201':
          description: User added to group successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  status:
                    type: string
                    description: Operation result status
                    example: "success"
                    minLength: 1
                    maxLength: 16
                    pattern: '^[a-zA-Z]+$'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied (not a group member)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/members/me:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    delete:
      tags: ["group"]
      summary: Leave a group
      description: Users have the option to leave a group at any time.
      operationId: leaveGroup
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Left the group successfully
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found or not a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/name:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    put:
      tags: ["group"]
      summary: Set the group name
      description: Update the name of a group.
      operationId: setGroupName
      security:
        - bearerAuth: []
      requestBody:
        description: The new group name string
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Rename request
              properties:
                name:
                  type: string
                  description: New name for the group
                  example: "New Group Name"
                  minLength: 1
                  maxLength: 64
                  pattern: '^[a-zA-Z0-9 _-]+$'
              required:
                - name
      responses:
        '200':
          description: Group name updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  status:
                    type: string
                    description: Operation result status
                    example: "success"
                    minLength: 1
                    maxLength: 16
                    pattern: '^[a-zA-Z]+$'
        '400':
          description: Invalid name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied (not a member)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/photo:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    put:
      tags: ["group"]
      summary: Set the group photo
      description: Upload or update the group photo.
      operationId: setGroupPhoto
      security:
        - bearerAuth: []
      requestBody:
        description: Binary image file
        required: true
        content:
          image/*:
            schema:
              type: string
              format: binary
              description: Image file data
      responses:
        '200':
          description: Group photo updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  status:
                    type: string
                    description: Operation result status
                    example: "success"
                    minLength: 1
                    maxLength: 16
                    pattern: '^[a-zA-Z]+$'
        '400':
          description: Invalid photo format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied (not a member)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'